+++ packages/app-project/src/screens/SubjectTalkPage/index.js
@@ -1,1 +1,1 @@
-export { default } from './SubjectTalkPage'
+export { default } from './SubjectTalkPageConnector'